+++ app/utils/getDateFormat.tsx
@@ -4,7 +4,7 @@ const getDateFormat = (date: string | Date) => {
   const year = formatter.getFullYear();
   const month = formatter.getMonth() + 1;
   const day = formatter.getDate();
-  return `${year}.${month >= 10 ? month : '0' + month}.${day}`;
+  return `${year}.${month >= 10 ? month : '0' + month}.${day >= 10 ? day : '0' + day}`;
 };
 
 export default getDateFormat;
